refactor: изменана структура серверного конфига
