+++ packages/server/index.tsx
@@ -15,5 +15,5 @@ const port = Number(process.env.SERVER_PORT) || 3001
 app.get('/*', renderTemplate)
 
 app.listen(port, () => {
-  console.log(`  ➜ 🎸 Server is listening on port: ${port}`)
+  console.log(`  ➜ 🎸 http://localhost:${port}`)
 })
