+++ packages/server/webpack.config.ts
@@ -1,52 +1,3 @@
-import type { Configuration } from 'webpack'
-import path from 'path'
+import { serverConfig } from './webpack'
 
-const DIST_DIR = path.join(__dirname, 'dist')
-
-const config: Configuration = {
-  target: 'node',
-  entry: './index.tsx',
-  module: {
-    rules: [
-      {
-        test: /\.(png|jpe?g|gif|svg)$/i,
-        use: [
-          {
-            loader: 'file-loader',
-          },
-        ],
-      },
-      {
-        test: /\.s[ac]ss$/i,
-        use: ['style-loader', 'css-loader', 'sass-loader'],
-      },
-      {
-        test: /\.(ts|tsx)$/,
-        exclude: /(node_modules)/,
-        use: {
-          loader: 'babel-loader',
-          options: {
-            presets: [
-              '@babel/env',
-              '@babel/preset-react',
-              '@babel/preset-typescript',
-            ],
-          },
-        },
-      },
-    ],
-  },
-  output: {
-    filename: 'index.js',
-    path: DIST_DIR,
-  },
-  resolve: {
-    modules: ['src', 'node_modules'],
-    extensions: ['*', '.js', '.jsx', '.json', '.ts', '.tsx'],
-    alias: {
-      '@': path.resolve(__dirname, '../client/src'),
-    },
-  },
-}
-
-export default config
+module.exports = [serverConfig]
