+++ packages/server/webpack/index.ts
@@ -1,0 +1,1 @@
+export { default as serverConfig } from './server.config'
