+++ packages/server/webpack/loaders/css/css.ts
@@ -1,0 +1,10 @@
+export default {
+  client: {
+    test: /\.s[ac]ss$/i,
+    use: ['style-loader', 'css-loader', 'sass-loader'],
+  },
+  server: {
+    test: /\.s[ac]ss$/i,
+    use: ['style-loader', 'css-loader', 'sass-loader'],
+  },
+}
