+++ packages/server/webpack/loaders/css/index.ts
@@ -1,0 +1,1 @@
+export { default as cssLoader } from './css'
