+++ packages/server/webpack/loaders/file/file.ts
@@ -1,0 +1,18 @@
+export default {
+  client: {
+    test: /\.(png|jpe?g|gif|svg)$/i,
+    use: [
+      {
+        loader: 'file-loader',
+      },
+    ],
+  },
+  server: {
+    test: /\.(png|jpe?g|gif|svg)$/i,
+    use: [
+      {
+        loader: 'file-loader',
+      },
+    ],
+  },
+}
