+++ packages/server/webpack/loaders/file/index.ts
@@ -1,0 +1,1 @@
+export { default as fileLoader } from './file'
