+++ packages/server/webpack/loaders/ts/index.ts
@@ -1,0 +1,1 @@
+export { default as tsPresets } from './ts'
