+++ packages/server/webpack/loaders/ts/ts.ts
@@ -1,0 +1,17 @@
+export default {
+  client: {},
+  server: {
+    test: /\.(ts|tsx)$/,
+    exclude: /(node_modules)/,
+    use: {
+      loader: 'babel-loader',
+      options: {
+        presets: [
+          '@babel/env',
+          '@babel/preset-react',
+          '@babel/preset-typescript',
+        ],
+      },
+    },
+  },
+}
